requirements.txt
# Sistema de Gestão de Dados do Tribunal
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.12.2
snowflake-connector-python==3.6.0
cryptography==41.0.7
python-dotenv==1.0.0
Pillow==10.1.0
cerberus==1.3.5
bleach==6.1.0
gunicorn==21.2.0
python-magic==0.4.27
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
structlog==23.2.0
python-json-logger==2.0.7
redis==5.0.1
flask-caching==2.1.0